Safeguards against missing params, missing paper

+++ app/controllers/references_controller.rb
@@ -2,11 +2,29 @@ class ReferencesController < ApplicationController
   before_action :ensure_current_user
 
   def create
-    list = List.find(reference_params[:list_id])
+    begin
+      reference_params.require(%i{type paper_id list_id})
+    rescue ActionController::ParameterMissing => e
+      missing_params = Set.new e.param[:reference]
 
-    # Check if uuid. If not, then ask Pubmed.
+      if missing_params.intersect? Set[:type, :paper_id]
+        flash['alert'] = 'You must choose a paper identifier type and input an ID.'
+      else
+        flash['alert'] = 'You must choose a list to add to.'
+      end
+
+      redirect_to :back
+      return
+    end
+
+    list = List.find(reference_params[:list_id])
     identifier = reference_params[:paper_id]
-    paper = find_paper(identifier) || import_paper(identifier)
+    id_type = reference_params[:type]
+
+    unless (paper = find_or_import_paper(id_type, identifier))
+      flash['alert'] = 'Paper not found.'
+      return redirect_to(:back)
+    end
 
     if Reference.exists? list_id: list.id, paper_id: paper.id
       flash['notice'] = 'This paper has already been added to this list'
@@ -17,29 +35,32 @@ class ReferencesController < ApplicationController
   end
 
   def destroy
-    reference = Reference.find(reference_params[:id])
-    list = reference.list
-    reference.destroy
-    redirect_to list
+    if (reference = Reference.find_by(id: reference_params[:id]))
+      reference.destroy
+      redirect_to reference.list
+    else
+      redirect_to :back
+    end
   end
 
   private
     def reference_params
-      params.require(:reference).permit(:list_id, :paper_id, :id)
+      params.require(:reference).permit(:list_id, :paper_id, :id, :type)
     end
 
-    def find_paper identifier
-      if UUID_FORMAT =~ identifier
-        paper = Paper.find(identifier)
-      elsif /\// =~ identifier
+    def find_or_import_paper id_type, identifier
+      case id_type
+      when 'link'
+        paper = Paper.find_by(link: identifier)
+      when 'doi'
         paper = Paper.find_by(doi: identifier)
-      else
-        paper = Paper.find_by(pubmed_id: identifier)
+      when 'pubmed'
+        paper = Paper.find_by(pubmed_id: identifier) || import_pubmed_paper(identifier)
       end
       return paper
     end
 
-    def import_paper identifier
+    def import_pubmed_paper identifier
       pubmed = Pubmed.new
       data = pubmed.search(identifier)['result'][identifier]
